pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nanorange"
version = "0.1.0"
description = "Agentic Microscopy Image Analysis System"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "NanoRange Team"}
]
keywords = [
    "microscopy",
    "image-analysis",
    "ai",
    "agent",
    "pipeline",
    "google-adk",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "google-adk>=1.0.0",
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
nanorange = "nanorange.main:main"

[project.urls]
Homepage = "https://github.com/nanorange/nanorange"
Documentation = "https://nanorange.readthedocs.io"
Repository = "https://github.com/nanorange/nanorange"

[tool.setuptools.packages.find]
where = ["."]
include = ["nanorange*"]

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I", "N"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
